# Adapters/OpenApi/openapi.yaml
openapi: 3.0.4
info:
  title: CLI
  version: '1.0'
paths:
...
  /api/v1/sites:
    get:
      tags:
        - Sites
      operationId: GetAllSitesV1
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - Sites
      operationId: CreateSiteV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  '/api/v1/sites/{id}':
    get:
      tags:
        - Sites
      operationId: GetSiteByIdV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Sites
      operationId: UpdateSiteV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Sites
      operationId: DeleteSiteV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    SiteRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
      additionalProperties: false
    SiteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: Enter the JWT token
      scheme: Bearer
      bearerFormat: JWT
...